refactor to modern project structure
