+++ src/main/java/com/devonfw/demoquarkus/service/mapper/AnimalMapper.java
@@ -1,4 +1,4 @@
-package com.devonfw.demoquarkus.rest.v1.mapper;
+package com.devonfw.demoquarkus.service.mapper;
 
 import java.util.List;
 
@@ -6,8 +6,8 @@ import org.mapstruct.Mapper;
 import org.tkit.quarkus.rs.mappers.OffsetDateTimeMapper;
 
 import com.devonfw.demoquarkus.domain.model.Animal;
-import com.devonfw.demoquarkus.rest.v1.model.AnimalDto;
-import com.devonfw.demoquarkus.rest.v1.model.NewAnimalDto;
+import com.devonfw.demoquarkus.service.model.AnimalDto;
+import com.devonfw.demoquarkus.service.model.NewAnimalDto;
 
 //mapstruct will generate an impl class(CDI bean, see pom.xml) from this interface at compile time
 @Mapper(uses = OffsetDateTimeMapper.class)
